requirements.txt
Django~=5.2
djangorestframework~=3.15
djangorestframework-simplejwt~=5.3
django-filter~=24.3
django-cors-headers~=4.4
python-dotenv~=1.0
